requirements.txt
bcrypt==4.2.0
email-validator==2.0.0.post2
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.0.5
psycopg2==2.9.9
python-dotenv==0.21.1
SQLAlchemy==2.0.32
typing_extensions==4.7.1
Werkzeug==2.2.3